requirements.txt
openai-whisper==20231117
numpy>=1.24.3
pydub>=0.25.1
pytest>=7.4.0